Create RibbonCableStrainRelief.js
Basic strain relief shape for ribbon cable over a terminal block or other connector. 
If there is room, this allows the cable to be secured by simply epoxying this over the top and down the sides. 
The biggest problem with this is that the side sections with interfere with anything to the side. 
And it does require superglue or epoxy to attach this part to the connector.